minor fix in refactor
